requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
asyncpg
sqlmodel
pydantic-settings
pyjwt
pwdlib[argon2,bcrypt]
python-multipart
